refactor: rename MyAppProvider into MyAppInfo
